+++ src/LinqAn.Google/Linq/Resolution/ProjectionReader.cs
