+++ src/LinqAn.Google/Linq/Resolution/RecordEnumerator.cs
