+++ src/LinqAn.Google/Linq/Resolution/RecordReader.cs
